Split catalog source workflows
